Gestión de alumnos como escuela #71

+++ ClimbCodeDjango/climbcode/actors/views.py
@@ -3,9 +3,8 @@ from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 from django.http import HttpResponseRedirect, HttpRequest
 from django.shortcuts import render, get_object_or_404
 
-from actors.forms import EditTeacherForm, RegisterTeacherForm
-from actors.models import Teacher, School
-
+from actors.forms import EditTeacherForm, RegisterTeacherForm, EditStudentForm, RegisterStudentForm
+from actors.models import Teacher, School, Student
 
 # Create your views here.
 from subjects.models import Subject
@@ -15,7 +14,7 @@ def list_teachers(request):
     user = request.user
     try:
         school = School.objects.get(userAccount_id=user.id)
-        teacher_list_aux = Teacher.objects.filter(school_t=school.id)
+        teacher_list_aux = Teacher.objects.filter(school_t=school)
     except Exception as e:
         teacher_list_aux = School.objects.none()
 
@@ -23,23 +22,26 @@ def list_teachers(request):
     paginator = Paginator(teacher_list_aux, 6)
 
     try:
-        teacher_list = paginator.page(page)
+        teacher_list_aux = paginator.page(page)
     except PageNotAnInteger:
-        teacher_list = paginator.page(1)
+        teacher_list_aux = paginator.page(1)
     except EmptyPage:
-        teacher_list = paginator.page(paginator.num_pages)
+        teacher_list_aux = paginator.page(paginator.num_pages)
 
     data = {
-        'teacher_list': teacher_list,
+        'teacher_list': teacher_list_aux,
         'title': 'Listado de profesores'
     }
     return render(request, 'teachers/list.html', data)
 
 
+
 def delete_teacher(request, pk):
     teacher = get_object_or_404(Teacher, pk=pk)
-    teacher.delete()
-    return HttpResponseRedirect('/actors/teachers/list')
+    if request.method == 'POST':
+        teacher.delete()
+        return HttpResponseRedirect('/actors/teachers/list')
+    return render(request, 'teachers/delete.html', {'teacher':teacher})
 
 
 def edit_teacher(request, pk):
@@ -52,8 +54,8 @@ def edit_teacher(request, pk):
         if (form.is_valid()):
             user = teacher.userAccount
 
-            userAccount.username = form.cleaned_data["username"]
-            userAccount.password = form.cleaned_data["password"]
+            userAccount.username = user.username
+            userAccount.password = user.password
             userAccount.email = form.cleaned_data["email"]
             userAccount.first_name = form.cleaned_data["first_name"]
             userAccount.last_name = form.cleaned_data["last_name"]
@@ -108,8 +110,7 @@ def register_teacher(request):
 
             try:
                 school = School.objects.get(userAccount_id=current_school.id)
-                school_id = school.id
-                teacher = Teacher.objects.create(phone=phone, photo=photo, dni=dni, userAccount=userAccount, school_t=school_id)
+                teacher = Teacher.objects.create(phone=phone, photo=photo, dni=dni, userAccount=userAccount, school_t=school)
             except Exception as e:
                 teacher = Teacher.objects.create(phone=phone, photo=photo, dni=dni, userAccount=userAccount)
 
@@ -126,3 +127,119 @@ def register_teacher(request):
     }
 
     return render(request, 'teachers/register.html', data)
+
+def list_students(request):
+    user = request.user
+    try:
+        school = School.objects.get(userAccount_id=user.id)
+        student_list_aux = Student.objects.filter(school_s=school)
+    except Exception as e:
+        student_list_aux = School.objects.none()
+
+    page = request.GET.get('page', 1)
+    paginator = Paginator(student_list_aux, 6)
+
+    try:
+        student_list_aux = paginator.page(page)
+    except PageNotAnInteger:
+        student_list_aux = paginator.page(1)
+    except EmptyPage:
+        student_list_aux = paginator.page(paginator.num_pages)
+
+    data = {
+        'student_list': student_list_aux,
+        'title': 'Listado de estudiantes'
+    }
+    return render(request, 'students/list.html', data)
+
+def register_student(request):
+    current_school = request.user
+
+    if (request.method == 'POST'):
+        form = RegisterStudentForm(request.POST, request.FILES)
+        if (form.is_valid()):
+
+            username = form.cleaned_data["username"]
+            password = form.cleaned_data["password"]
+            email = form.cleaned_data["email"]
+            first_name = form.cleaned_data["first_name"]
+            last_name = form.cleaned_data["last_name"]
+
+            user = User.objects.create_user(username, email, password)
+            user.first_name = first_name
+            user.last_name = last_name
+            user.save()
+
+            # Creación del estudiante
+            phone = form.cleaned_data["phone"]
+            photo = form.cleaned_data["photo"]
+            dni = form.cleaned_data["dni"]
+
+            userAccount = user
+
+            try:
+                school = School.objects.get(userAccount_id=current_school.id)
+                student = Student.objects.create(phone=phone, photo=photo, dni=dni, userAccount=userAccount, school_s=school)
+            except Exception as e:
+                student = Student.objects.create(phone=phone, photo=photo, dni=dni, userAccount=userAccount)
+
+            student.save()
+
+            return HttpResponseRedirect('/actors/students/list')
+
+    else:
+        form = RegisterStudentForm()
+
+    data = {
+        'form': form,
+        'title': 'Registrar estudiante'
+    }
+
+    return render(request, 'students/register.html', data)
+
+
+def edit_student(request, pk):
+    assert isinstance(request, HttpRequest)
+    student = get_object_or_404(Student, pk=pk)
+    userAccount = get_object_or_404(User, pk=student.userAccount_id)
+
+    if (request.method == 'POST'):
+        form = EditStudentForm(request.POST, request.FILES)
+        if (form.is_valid()):
+            user = student.userAccount
+
+            userAccount.username = user.username
+            userAccount.password = user.password
+            userAccount.email = form.cleaned_data["email"]
+            userAccount.first_name = form.cleaned_data["first_name"]
+            userAccount.last_name = form.cleaned_data["last_name"]
+
+            userAccount.save()
+
+            student.phone = form.cleaned_data["phone"]
+            student.photo = form.cleaned_data["photo"]
+            student.dni = form.cleaned_data["dni"]
+
+            student.save()
+
+            return HttpResponseRedirect('/actors/students/list')
+
+    elif request.method == 'DELETE':
+        student.delete()
+    else:
+        form = EditStudentForm()
+
+    data = {
+        'form': form,
+        'student': student,
+        'title': 'Editar estudiante'
+    }
+
+    return render(request, 'students/edit.html', data)
+
+def delete_student(request, pk):
+    student = get_object_or_404(Student, pk=pk)
+    if request.method == 'POST':
+        student.delete()
+        return HttpResponseRedirect('/actors/students/list')
+    return render(request, 'students/delete.html', {'student':student})
